requirements.txt
numpy~=1.22.0
torch~=1.10.1
torchvision~=0.11.2
matplotlib~=3.5.1
Pillow~=9.0.0
requests~=2.27.1